Support of set of local dev properties has been added.
File 'application-dev.properties' can be added to the root of the
project and local dev properties can be defined in it. To make this file
copy (or use part of it as source) application-default.properties and
edit it. Git configured to ignore file application-dev.properties.

Example of a file defined local dev environment specific properties
# Profile name can be specified with property 'spring.profiles.active'
# e.g. mvn spring-boot:run --spring.profiles.active=dev
#